feat(model-validation): fix flake8 errors
