+++ tests/model_serving/model_runtime/model_validation/conftest.py
@@ -10,9 +10,7 @@ from ocp_resources.namespace import Namespace
 from ocp_resources.serving_runtime import ServingRuntime
 from ocp_resources.inference_service import InferenceService
 from pytest import FixtureRequest
-from pytest_testconfig import config as py_config
 from kubernetes.dynamic import DynamicClient
-from ocp_resources.cluster_service_version import ClusterServiceVersion
 from simple_logger.logger import get_logger
 from utilities.serving_runtime import ServingRuntimeFromTemplate
 from utilities.inference_utils import create_isvc
@@ -27,52 +25,12 @@ from syrupy.extensions.json import JSONSnapshotExtension
 LOGGER = get_logger(name=__name__)
 
 
-@pytest.fixture(scope="package")
-def fail_if_missing_dependent_operators(admin_client: DynamicClient) -> None:
-    if dependent_operators := py_config.get("dependent_operators"):
-        missing_operators: list[str] = []
-
-        for operator_name in dependent_operators.split(","):
-            csvs = list(
-                ClusterServiceVersion.get(
-                    dyn_client=admin_client,
-                    namespace=py_config["applications_namespace"],
-                )
-            )
-
-            LOGGER.info(f"Verifying if {operator_name} is installed")
-            for csv in csvs:
-                if csv.name.startswith(operator_name):
-                    if csv.status == csv.Status.SUCCEEDED:
-                        break
-
-                    else:
-                        missing_operators.append(
-                            f"Operator {operator_name} is installed but CSV is not in {csv.Status.SUCCEEDED} state"
-                        )
-
-            else:
-                missing_operators.append(f"{operator_name} is not installed")
-
-        if missing_operators:
-            pytest.fail(str(missing_operators))
-
-    else:
-        LOGGER.info("No dependent operators to verify")
-
-
-@pytest.fixture(scope="session")
-def skip_if_no_supported_accelerator_type(supported_accelerator_type: str) -> None:
-    if not supported_accelerator_type:
-        pytest.skip("Accelartor type is not provided,vLLM test can not be run on CPU")
-
-
 @pytest.fixture(scope="class")
 def vllm_model_car_inference_service(
     request: FixtureRequest,
     admin_client: DynamicClient,
     model_namespace: Namespace,
-    serving_runtime: ServingRuntime,
+    modelcar_serving_runtime: ServingRuntime,
     supported_accelerator_type: str,
     modelcar_image_uri: str,
 ) -> Generator[InferenceService, Any, Any]:
@@ -80,9 +38,9 @@ def vllm_model_car_inference_service(
         "client": admin_client,
         "name": request.param["name"],
         "namespace": model_namespace.name,
-        "runtime": serving_runtime.name,
+        "runtime": modelcar_serving_runtime.name,
         "storage_uri": modelcar_image_uri,
-        "model_format": serving_runtime.instance.spec.supportedModelFormats[0].name,
+        "model_format": modelcar_serving_runtime.instance.spec.supportedModelFormats[0].name,
         "deployment_mode": request.param.get("deployment_mode", KServeDeploymentType.SERVERLESS),
         "image_pull_secrets": [ORIGINAL_PULL_SECRET],
     }
@@ -119,7 +77,7 @@ def vllm_model_car_inference_service(
 
 
 @pytest.fixture(scope="class")
-def serving_runtime(
+def modelcar_serving_runtime(
     request: FixtureRequest,
     admin_client: DynamicClient,
     model_namespace: Namespace,
